fModif ProfilAndMaps: création du composant MapButton pour gérer l’état et l’animation du bouton
